requirements.txt
gunicorn==21.2.0
Flask==2.3.3
Flask-Cors==3.0.10
SQLAlchemy>=2.0.31
APScheduler==3.10.1
python-dateutil==2.8.2
tldextract==3.4.0
whois==0.9.27
dnspython==2.3.0
requests==2.31.0
itsdangerous==2.1.2
python-dotenv==1.0.0
PyJWT==2.8.0
